Fix warnings for eth_hw_addr_set and event->bssid
esp_set_mac_address wont change the mac address of the device, so
we can remove it acording to the kernel documentation.

!event->bssid will always be true assuming that event is not NULL,
so we can remove the check.

Use eth_hw_addr_set instead of ether_addr_copy to set the mac address
of the device for >= 5.15.
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=f3956ebb3bf06ab2266ad5ee2214aed46405810c

request->ssids can be NULL, so we need to check it before using it.

Signed-off-by: Jesse Taube <Mr.Bossman075@gmail.com>