FG: Use explicit macro for MCU ports.
The current FG ctrl_api relies "STM32F469xx" macro to be defined for
MCU hosts without unistd and socket APIs. As this macro is implicitly
defined by STM projects when the target MCU is the same as the
discovery board (STM32F469I), this could be a issue if a different
MCU platform is used.
This patch make this selection clear by using a seperate macro defined in
"platform_wrapper.h", which can be defined by the user as required.

Signed-off-by: Yilin Sun <imi415@imi.moe>