Return err code when a Ringbuffer is full or inaccessible
esp_rb_write_by_kernel() returns the number of bytes written; and
it is called repeatedly until the entire length of the packet has
been transferred. In the case of a full/inaccessible ringbuffer, it
was returning zero, causing the caller to invoke the function in
a tight loop. Fixed to return -EFAULT.