bugfix(Windows) Fix .project file generation
On Windows, in the generated .project file, STM32CubeIDE expects the
variable with the name CODE_BASE to use forward slash, '/', as the
separator to pick up the code base correctly. For example:

<value>file:/c:/esp-hosted/esp_hosted_fg</value>

Currently, the batch file generates this incorrect value:

<value>file:c:\esp-hosted\esp_hosted_fg</value>

which uses the standard Windows back slash, '\', as the separator and
does not prefix the drive 'c:' with a slash.

This fix modifies the batch file to generate the correct value.

Following variations were also tested, and did not work with
STM32CubeIDE:

<value>file:\c:\esp-hosted\esp_hosted_fg</value>

<value>file:/c:\esp-hosted\esp_hosted_fg</value>