bugfix(slave/rcp) Fix bugs in slave esp_wifi_get_config, host rpc handling
On slave, fixed memory violations when processing result of
esp_wifi_get_config()

On host, added code to handle case where a sync RPC request fails to
be sent. An error response is generated and returned, so the RPC does
not timeout waiting for a response that does not come (as no request
was sent).

On host, added checks for valid response data before copying data.