spi_slave_api.c: Allow to use GPIO 32 and 33 for interface signals.
Using consistently the IDF API instead of direct port access. There
is a small performace drop compared to direct port access for setting
the pin, but that can be accepted in favoer of consistence.

GPIO33 for instance is used by Adafruit and NINA modules for the
handshake signal, and these modules cannot be rewired.

Signed-off-by: robert-hh <robert@hammelrath.com>