Change scan type from passive to active
I ran into a situation where I was unable to connect to a particular AP
and tracked it down to the scan type being passive vs an active one.

Perhaps we could tune the beacon frequency on this AP, but I figured
that I would bring up the matter with you guys.  :-)

Even if this is not the desired behavior and you wish to keep the
passive scan in place, the enum should be used instead of '1'.